Handle packages name and folders
