+++ templates/spring-boot-camel-api-consumer/skeleton/pom.xml
@@ -112,9 +112,9 @@
                                 <version>${project.version}</version>
                                 <dateLibrary>java8</dateLibrary>
                                 <sourceFolder>src/main/java</sourceFolder>
-                                <basePackage>${{ values.java_package_domain }}.${{ values.java_group_name }}.${{ values.java_package_name }}</basePackage>
-                                <modelPackage>${{ values.java_package_domain }}.${{ values.java_group_name }}.${{ values.java_package_name }}.model</modelPackage>
-                                <apiPackage>${{ values.java_package_domain }}.${{ values.java_group_name }}.${{ values.java_package_name }}.api</apiPackage>
+                                <basePackage>${{ values.java_package_name }}</basePackage>
+                                <modelPackage>${{ values.java_package_name }}.model</modelPackage>
+                                <apiPackage>${{ values.java_package_name }}.api</apiPackage>
                                 <documentationProvider>none</documentationProvider>
                                 <useSpringController>false</useSpringController>
                                 <interfaceOnly>true</interfaceOnly>
